Add Entry Class, Yornal filter, entries, last methods
Entry Class allows for easier managing of entries in a Yornal.
Implements Comparable interface for quick comparison/sort.
Has easy conversion from and to Time objects.

Yornal last() method aware of anchor point from `from`.
Anchor point = time from which to calculate the past.
If from (entry array) is given, the anchor is not Time.now,
but the last element in the array.
Useful for interop between --last and --from command line switches.